♻️ refactor(db): update schema barrel file and fix test imports
- Update db schema index.ts to properly export auth schema types
- Fix question routes integration test to use direct UserId import
